requirements.txt
streamlit
pandas
langchain
langchain-google-genai
langchain-experimental
google-generativeai
matplotlib
seaborn
tabulate